refactor: dummy image
