+++ src/pages/deck-page/deck-page.tsx
@@ -2,7 +2,7 @@ import { ChangeEvent, useState } from 'react'
 import { Link, useParams } from 'react-router-dom'
 
 import { ArrowBackOutline } from '@/assets/icons'
-import myImage from '@/assets/webp/react-logo.webp'
+import myImage from '@/assets/webp/dummy-cover.webp'
 import { AddNewCardDialogForm, AddNewDeckDialogForm, DeleteDialogForm } from '@/components/forms'
 import { DeckTable, DeckTitle } from '@/components/ui/layout-components'
 import { Button, Progress, TextField } from '@/components/ui/primitives'
